MPIParticleInCell

MPIParticleInCell.AbstractGrid

MPIParticleInCell.Field

MPIParticleInCell.UniformGrid

MPIParticleInCell.communicate_halo!

MPIParticleInCell.compute_proc_bounds

MPIParticleInCell.compute_proc_indices

MPIParticleInCell.compute_proc_rank

MPIParticleInCell.compute_processor_dimensions

MPIParticleInCell.decompose_domain

MPIParticleInCell.gather_field_on_root!

MPIParticleInCell.AbstractGrid — Type

Stores global and per-rank information about a grid imposed over a simulation domain.

MPIParticleInCell.Field — Type

Field(grid, name, data_type, full_field_on_root)

Creates a distributed array of type data_type with dimensions specified by grid. If full_field_on_root is true, then the root node allocates memory the entire array, although in normal operation it only utilizes the portion of memory that it is assigned to update. Obviously, care must be taken to not use this option for large simulations.

MPIParticleInCell.UniformGrid — Type

UniformGrid(lower_dims, upper_dims, num_cells, num_guard_cells, periodic)

Stores global and per-rank information about a uniform N-dimensional spatial grid.

MPIParticleInCell.communicate_halo! — Method

communicate_halo!(f::Field, comm)

Communicates the values of each processors guard cells to its neighbors.

MPIParticleInCell.compute_proc_bounds — Method

compute_proc_bounds(num_cells, proc_dims, proc)

Computes the range of cells that proc is responsible for.

MPIParticleInCell.compute_proc_indices — Method

compute_proc_indices(proc_dims, proc)

Computes the unique indices in the processor grid for proc.

MPIParticleInCell.compute_proc_rank — Method

compute_proc_rank(proc_dims, indices)

Computes the rank associated with the processor at indices.

MPIParticleInCell.compute_processor_dimensions — Method

compute_processor_dimensions(num_cells::NTuple{N}, num_procs)

Given the desired number of cells in each dimension, and a number of processors, returns the number of subdivisions along each dimension that minimize the amount of inter-rank communication.

MPIParticleInCell.decompose_domain — Method

decompose_domain(num_cells::NTuple{N}, num_procs[, proc])

Given a target number of pieces num_procs and an N-cube simulation domain with num_cells along each dimension, computes a decomposition of the domain. If proc is specified, returns a pair of Tuples corresponding to the lower and upper bounds of the cells assigned to proc. Otherwise, returns a pair of vectors of the lower and upper bounds for each processor.

NOTE: proc is zero indexed.

MPIParticleInCell.gather_field_on_root! — Method

gather_field_on_root!(f::Field)

Gathers the entire field f on the root node. Can only be used if f was created with full_field_on_root = true.